+++ NEWS.md
@@ -1,3 +1,21 @@
+# datapasta 3.1.0
+* Exported `_format` functions
+* Adds `dt_paste` function for pasting as `data.table` (Thanks @jonocarroll, #72, closes #70)
+* Row names are kept in data.frames and data.tables (Thanks @sowla)
+* Column names that are invalid are now handled with backticks (Thanks @sharlagelfand)
+* Fixes issue with commas inside character vectors getting wrapped on
+* data.frame (and data.table) print is much prettier and robust with all args and cols aligned on '='
+* zero row tibbles are supported with a fall-back to a tibble::tibble() call
+* all _construct functions now return input visibly
+
+# datapasta 3.0.0 'Colander Helmet'
+
+* When pasting from clipboard it now attempts to guess if there is no header row, in the case where the clipboard is all data. If you're lucky it will create a default header for you when pasting (V1, V2, V3 etc.).
+* `dpasta()` will now handle tribbles with R classes that cannot be represented in tribble form. It falls back to their character representation. This works well for things like dates.
+* New addin: 'Fiddle Selection'. This is a kind of magic wand that can be waved over RStudio editor selections to: Reflow messy tribble and data.frame definitions, create `c()` expressions from raw data, and pivot `c()` exprs between vertical and horizontal format.
+* New addin: 'Toggle Vector Quotes'. Given a horizontal or vertical `c()` expr, it will toggle all elements between quoted and bare format.
+* Complies with new CRAN policy on clipboard use. You cannot write to the clipboard in non-interactive sessions with `dmdclip()` - Why would you?. Tests containing clipboard use are skipped on CI and CRAN.
+
 # datapasta 2.0.1
 
 * Added a trailing newline after all pastes, this works much nicer for console output.